pgn handlers moved into a hook
